+++ src/utils/formatting/content_split/normal.rs
@@ -49,25 +49,3 @@ pub fn split_long_word(allowed_width: usize, word: &str) -> (String, String) {
     let remaining = char_iter.collect();
     (parts, remaining)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_split_long_word() {
-        let emoji = "🙂‍↕️"; // U+1F642 U+200D U+2195 U+FE0F head shaking vertically
-        assert_eq!(emoji.len(), 13);
-        assert_eq!(emoji.chars().count(), 4);
-        assert_eq!(emoji.width(), 2);
-
-        let (word, remaining) = split_long_word(emoji.width(), emoji);
-
-        assert_eq!(word, "\u{1F642}\u{200D}\u{2195}\u{FE0F}");
-        assert_eq!(word.len(), 13);
-        assert_eq!(word.chars().count(), 4);
-        assert_eq!(word.width(), 2);
-
-        assert!(remaining.is_empty());
-    }
-}
